Rename folder and simplify `lib.rs` Rename the example folder from `basic_diesel` to `diesel` Use `embed_migrations!()` to simplify the logic in `lib.rs`
