--- examples/basic_diesel/.env
@@ -1,1 +1,0 @@
-DATABASE_URL=.posts.db
