+++ examples/diesel/Cargo.toml
@@ -16,6 +16,7 @@ futures = "~0.1.11"
 mime = "0.3"
 log = "0.3"
 diesel = { version = "1.0.0", features = ["sqlite"] }
+diesel_migrations = { version = "1.0.0", features = ["sqlite"] }
 dotenv = "0.9.0"
 r2d2 = "0.8.2"
 r2d2-diesel = "1.0.0"
