+++ examples/diesel/src/models.rs
