+++ examples/diesel/src/schema.rs
